First working version.
At this point we can 'execute' DNA to generate RNA and then 'build' that
RNA to generate an image.  The functions are tested and should work but
right now it takes five minutes to fill an empty bitmap.  That is way
too long and I'll need to speed up fill() before really solving this
problem.