Combine dna.{cpp,hpp} into execute.{cpp,hpp}
Now there is one header and one source file for each of the programs
specified in the task documentation.  (Originally I thought there would
be a need to have an object that was common between the programs but
processing the DNA is nothing like processing the RNA even though they
both use the same character sequence.